queen_V1.1_patch1.5/infra/dockerfiles/queen_core/requirements.txt
redis>=5.0,<6.0
requests>=2.31,<3.0
pyyaml>=6.0,<7.0
